# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards REST API
  version: 1.0.0
  description: API для управления банковскими картами, пользователями и переводами.

servers:
  - url: http://localhost:8080
    description: Local server

tags:
  - name: Auth
  - name: Users
  - name: Cards

paths:
  /api/auth:
    post:
      tags: [Auth]
      summary: Войти в систему
      description: Проверяет логин/пароль и возвращает JWT токен.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDto'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponseDto'
        '401':
          description: Неверный пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/users:
    post:
      tags: [Users]
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationDto'
      responses:
        '201':
          description: Пользователь создан
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Ошибка валидации/бизнес-ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
    get:
      tags: [Users]
      summary: Получить всех пользователей
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/users/{userId}/role:
    patch:
      tags: [Users]
      summary: Изменить роль пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: role
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Роль обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/users/{userId}:
    delete:
      tags: [Users]
      summary: Удалить пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Удален
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards:
    post:
      tags: [Cards]
      summary: Выпуск карты
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Карта создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards/all:
    get:
      tags: [Cards]
      summary: Все карты (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/DirectionParam'
      responses:
        '200':
          description: Страница карт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCardResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards/user/{userId}:
    get:
      tags: [Cards]
      summary: Карты пользователя (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/DirectionParam'
      responses:
        '200':
          description: Страница карт пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCardResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards/me:
    get:
      tags: [Cards]
      summary: Свои карты (USER)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/DirectionParam'
      responses:
        '200':
          description: Страница карт текущего пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCardResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards/{cardId}/block-request:
    patch:
      tags: [Cards]
      summary: Запрос блокировки карты (USER)
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта заблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '400':
          description: Бизнес-ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards/{cardId}/block:
    patch:
      tags: [Cards]
      summary: Блокировка карты (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта заблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards/{cardId}/activate:
    patch:
      tags: [Cards]
      summary: Активация карты (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта активирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards/{cardId}:
    delete:
      tags: [Cards]
      summary: Удаление карты (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Карта удалена
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

  /api/cards/transfer:
    post:
      tags: [Cards]
      summary: Перевод между своими картами (USER)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequestDto'
      responses:
        '200':
          description: Перевод выполнен
        '400':
          description: Ошибка валидации/бизнес-ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 10
    DirectionParam:
      name: direction
      in: query
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: asc

  schemas:
    Role:
      type: string
      enum: [USER, ADMIN]

    CardStatus:
      type: string
      enum: [ACTIVE, BLOCKED, EXPIRED]

    UserRegistrationDto:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          minLength: 6

    UserLoginDto:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    JwtResponseDto:
      type: object
      properties:
        token:
          type: string

    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'

    TransferRequestDto:
      type: object
      required: [fromCardId, toCardNumber, amount]
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardNumber:
          type: string
        amount:
          type: number
          format: double
          minimum: 0.01

    CardResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumber:
          type: string
          description: Замаскированный номер карты
          example: "**** **** **** 1234"
        expirationDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          format: double

    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string

    PageCardResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponseDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean